==> Differential_CpG_Analysis/5.HOMER_Motif_Finder.R <==
findMotifsGenome.pl Embryo_Larva_Both.txt \
~/Nvit_psr_1.1/GCF_009193385.2_Nvit_psr_1.1_genomic.fa \new_Embryo_Larva/ -cpg

#Where Embryo_Larva_Both.txt is a file created with the DMS' with each differentially methylated site being added to the + and - strand as DMS' in this analysis are destranded
